:bento: Chore: 주간 랭킹 api data class 정의

+++ app/src/main/java/com/example/snack4diet/api/weeklyRank/Data.kt
@@ -1,12 +1,12 @@
 package com.example.snack4diet.api.weeklyRank
 
 data class Data(
-    val calorieScore: Int,
-    val carbohydrateScore: Int,
-    val fatScore: Int,
+    val calorieScore: Double,
+    val carbohydrateScore: Double,
+    val fatScore: Double,
     val image: String,
     val name: String,
-    val proteinScore: Int,
-    val totalScore: Int,
-    val userId: Int
+    val proteinScore: Double,
+    val totalScore: Double,
+    val userId: Long
 )
